# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Routeyou events API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/events/{event}/invite": {
            "post": {
                "tags": [
                    "Mail"
                ],
                "summary": "Invite users to an event",
                "operationId": "ea65274a61b285eba372dab71cfcc87c",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID of the event to invite users to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "List of emails to invite",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "followers"
                                ],
                                "properties": {
                                    "followers": {
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "emails": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "test1@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "email": "test1@example.com"
                                            },
                                            {
                                                "email": "test2@example.com"
                                            },
                                            {
                                                "email": "test3@example.com"
                                            }
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The invitations have been sent successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "followers": [
                                                    "The followers field is required."
                                                ],
                                                "emails": [
                                                    "The emails must be an array."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Token": [
                            "ORGANISER"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}